requirements.txt
fastapi==0.105.0
httpx==0.26.0
numpy==1.26.2
orjson==3.9.10
pytest==7.4.3
python-dotenv==1.0.0
uvicorn==0.24.0.post1